backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic-settings==2.1.0
python-multipart==0.0.9
alembic==1.12.1
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.26.0
tzdata==2024.1
